Update 36_Double-base-palindromes.js
There is a bug in the JavaScript code due to the way the variables n and k are extracted and used. Both are initially extracted as strings from the input, but they should be converted to numbers before using them in calculations or loops. Specifically:

n is used in a loop, but it's a string, so it needs to be converted to a number.
k is used in the toString(k) method to convert numbers to base k, but it also needs to be converted to a number.